build: bump up version for testing
The serious problem is the intent-filter-verification won't be executed
in app re-installation, even if we bump up app version.

"Maybe", maybe it is due to adb command, and it might be re-executed if
the app is installed via Google-play-store. But we don't know until to
give a try on production site.